Addition to controllers - finishing CRUD
All existing tables now have full CRUD functionality + local endpoints.

+++ Experling-API/DataAccess/Models/TicketModel.cs
@@ -13,7 +13,7 @@ namespace DataAccess.Models
         public int Eventid { get; set; }
         public string CustomerName { get; set; }
         public string CustomerEmail { get; set; }
-        public int AgeCheck { get; set; }
+        public byte AgeCheck { get; set; }
         public string GUid { get; set; }
     }
 }
